Render environment changes.
Nothing gets rendered unless there is an event that would update the content on screen.
Instead of clearing the screen we move the cursor to the top and clear everything after it.
That way the we don't shift the console output down a page every frame, so it's similar
to how nano works visually.