Fixed tty flicker and reformat
The tty does not vsync in anyway and so one should not clear anything on it.
If you clear than the clear can appear visually to the user before you draw the
replacement text. The solution is to draw directly over the old text and put
spaces where you want to remove.